moved navChoice model to core folder, removed user model (unused)
