--- client/src/app/_models/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-  email?: string;
-  password?: string;
-  _id?: string;
-}
